just keep one kind of executer
